Integer datatype for output names (#159)
* Integer datatype for output names

Added specifying integer datatype for output file name

* Output name datatype update

Updated the datatype specification for output image names inside the function file.

+++ scripts/align_z_focus.py
@@ -61,7 +61,7 @@ def get_sms_im_name(time_idx=None,
     if pos_idx is not None:
         im_name += "_p" + str(pos_idx).zfill(int2str_len)
     if slice_idx is not None:
-        im_name += "_z" + str(slice_idx).zfill(int2str_len)
+        im_name += "_z" + str(slice_idx.astype('int64')).zfill(int2str_len)
     if extra_field is not None:
         im_name += "_" + extra_field
     im_name += ext
